Remove ReaderTabItem
